+++ src/main/java/com/app/dao/EventDao.java
@@ -6,8 +6,8 @@ import com.app.model.Event;
 import com.app.model.Payment;
 
 public interface EventDao {
-	public Event findById(int id) throws Exception;
-	public void deleteById(int id) throws Exception;
+	public Event findById(Long id) throws Exception;
+	public void deleteById(Long id) throws Exception;
 	public void insert(Event event) throws Exception;
 	public void update(Event event) throws Exception;
 	public List<Event> findAll() throws Exception;
