+++ src/main/java/com/app/service/EventService.java
@@ -5,8 +5,8 @@ import java.util.List;
 import com.app.model.Event;
 
 public interface EventService {
-	public Event findById(int id) throws Exception;
-	public void deleteById(int id) throws Exception;
+	public Event findById(Long id) throws Exception;
+	public void deleteById(Long id) throws Exception;
 	public void insert(Event event) throws Exception;
 	public void update(Event event) throws Exception;
 	public List<Event> findAll() throws Exception;
